fixed most exceptions
