+++ Plugin/Patches/__System/DateTime.cs
@@ -10,7 +10,7 @@ class get_Now
 {
     static Exception Cleanup(MethodBase original, Exception ex)
     {
-        return AuxilaryHelper.Cleanup_IgnoreException(original, ex);
+        return PatcherHelper.Cleanup_IgnoreException(original, ex);
     }
 
     static bool Prefix(ref DateTime __result)
