+++ Plugin/Patches/__System/Environment.cs
@@ -10,7 +10,7 @@ class get_TickCount
 {
     static Exception Cleanup(MethodBase original, Exception ex)
     {
-        return AuxilaryHelper.Cleanup_IgnoreException(original, ex);
+        return PatcherHelper.Cleanup_IgnoreException(original, ex);
     }
 
     static bool Prefix(ref int __result)
