+++ Plugin/Patches/__UnityEngine/Resources.cs
@@ -19,7 +19,7 @@ class LoadAsyncInternal
 {
     static Exception Cleanup(MethodBase original, Exception ex)
     {
-        return AuxilaryHelper.Cleanup_IgnoreException(original, ex);
+        return Patches.PatcherHelper.Cleanup_IgnoreException(original, ex);
     }
 
     static bool Prefix(string path, Type type, ref object __result)
