+++ Plugin/Patches/__UnityEngine/__EventSystems/EventSystem.cs
@@ -22,7 +22,7 @@ class isFocusedGetter
 
     static Exception Cleanup(MethodBase original, Exception ex)
     {
-        return AuxilaryHelper.Cleanup_IgnoreException(original, ex);
+        return Patches.PatcherHelper.Cleanup_IgnoreException(original, ex);
     }
 
     static bool Prefix(ref bool __result)
